+++ src/Reid/V20190928/ReidApiResolver.php
@@ -12,9 +12,11 @@ use AlibabaCloud\Client\Resolver\ApiResolver;
  * @method DescribeDevices describeDevices(array $options = [])
  * @method DescribeHeatMap describeHeatMap(array $options = [])
  * @method DescribeImageUrls describeImageUrls(array $options = [])
+ * @method DescribeIpcLiveAddress describeIpcLiveAddress(array $options = [])
  * @method DescribeOverviewData describeOverviewData(array $options = [])
  * @method ImportSpecialPersonnel importSpecialPersonnel(array $options = [])
  * @method ListActionData listActionData(array $options = [])
+ * @method ListDevicesImages listDevicesImages(array $options = [])
  * @method ListEmap listEmap(array $options = [])
  * @method ListLocation listLocation(array $options = [])
  * @method ListMaskDetectionResults listMaskDetectionResults(array $options = [])
@@ -431,6 +433,40 @@ class DescribeImageUrls extends Rpc
 }
 
 /**
+ * @method string getIpcId()
+ * @method string getStoreId()
+ */
+class DescribeIpcLiveAddress extends Rpc
+{
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withIpcId($value)
+    {
+        $this->data['IpcId'] = $value;
+        $this->options['form_params']['IpcId'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withStoreId($value)
+    {
+        $this->data['StoreId'] = $value;
+        $this->options['form_params']['StoreId'] = $value;
+
+        return $this;
+    }
+}
+
+/**
  * @method string getDate()
  * @method string getStoreIds()
  */
@@ -667,6 +703,40 @@ class ListActionData extends Rpc
     {
         $this->data['FilterInvalidData'] = $value;
         $this->options['form_params']['FilterInvalidData'] = $value;
+
+        return $this;
+    }
+}
+
+/**
+ * @method string getIpcIdList()
+ * @method string getStoreId()
+ */
+class ListDevicesImages extends Rpc
+{
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withIpcIdList($value)
+    {
+        $this->data['IpcIdList'] = $value;
+        $this->options['form_params']['IpcIdList'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withStoreId($value)
+    {
+        $this->data['StoreId'] = $value;
+        $this->options['form_params']['StoreId'] = $value;
 
         return $this;
     }